requirements.txt
Django>=5.0,>4.0,>3.0
google-generativeai>=0.7.2
django-ckeditor-5==0.2.13
dj-database-url==2.2.0
Pillow>=9.0
ipython==8.26.0
python-dotenv==1.0.1
psycopg2>=2.9.9
django-rq>=0.2.1
redis
json5